src/main/java/org/oandx/rest/request/ChoiceRequest.java
package org.oandx.rest.request;

public class ChoiceRequest {

    private int x;
    private int y;
    private int character;
    private int boardId;

    public int getX() {
        return x;
    }

    public void setX(int x) {
        this.x = x;
    }

    public int getY() {
        return y;
    }

    public void setY(int y) {
        this.y = y;
    }

    public int getCharacter() {
        return character;
    }

    public void setCharacter(int character) {
        this.character = character;
    }

    public int getBoardId() {
        return boardId;
    }

    public void setBoardId(int boardId) {
        this.boardId = boardId;
    }
}
